refactor(unity-daemon): inline container wiring into bootstrap
Why:
- Daemon bootstrap composition was spread across dedicated DI helper/factory types without adding runtime value.

What Changed:
- Moved Unity daemon service registration and container construction directly into UnityDaemonBootstrap.
- Removed Composition-layer DI helper/factory types and related Unity meta files.

Impact:
- Bootstrap entrypoint now owns container wiring in one place.
- Composition abstraction layer for daemon bootstrap is removed.
