--- src/Ucli.Unity/Assets/MackySoft/MackySoft.Ucli.Unity/Editor/Ipc/Composition.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 38443bd46b94d46aa83771fb7f033984
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
