--- src/Ucli.Unity/Assets/MackySoft/MackySoft.Ucli.Unity/Editor/Ipc/Composition/IUnityDaemonServiceProviderFactory.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6790dd8d2a0d04c23908511aa8f8203d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
