--- src/Ucli.Unity/Assets/MackySoft/MackySoft.Ucli.Unity/Editor/Ipc/Composition/UnityDaemonServiceCollectionExtensions.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f2bdb61bf3dc14dab91849add85072fb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
