--- src/Ucli.Unity/Assets/MackySoft/MackySoft.Ucli.Unity/Editor/Ipc/Composition/UnityDaemonServiceProviderFactory.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 94129fc5d48ca44cfafbdf64870388f2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
